+++ HarfangLab/harfanglab/base.py
@@ -1,8 +1,11 @@
 import re
 from functools import cached_property
 from posixpath import join as urljoin
+from typing import Any
 
+import orjson
 from sekoia_automation.action import GenericAPIAction
+from sekoia_automation.exceptions import MissingActionArgumentError, MissingActionArgumentFileError
 
 
 class HarfanglabAction(GenericAPIAction):
@@ -38,3 +41,27 @@ class HarfanglabAction(GenericAPIAction):
 
             path += f"?{'&'.join(query_arguments)}"
         return path
+
+    def json_argument(self, name: str, arguments: dict, required: bool = True) -> Any:
+        """Get a JSON Argument by direct reference of by reading a file.
+
+        If `name` is inside arguments, returns the value.
+        If `name`_path is inside arguments, returns the content of the file
+        """
+        # @todo doesn't work with booleans by default - fix the error in Sekoia SDK
+        if arguments.get(name, None) is not None:
+            self._result_as_file = False
+            return arguments[name]
+
+        elif f"{name}_path" in arguments:
+            self._result_as_file = True
+            filepath = self.data_path.joinpath(arguments[f"{name}_path"])
+            if not filepath.is_file():
+                raise MissingActionArgumentFileError(filepath)
+
+            with filepath.open("r") as f:
+                return orjson.loads(f.read().encode("utf-8"))
+
+        else:
+            if required:
+                raise MissingActionArgumentError(name)